requirements.txt
PyInquirer==1.0.3
Click==7.1.2
tinydb==4.2.0
gitpython==3.1.9